7 ajax

+++ app/controllers/answers_controller.rb
@@ -1,26 +1,41 @@
 class AnswersController < ApplicationController
-
   before_action :authenticate_user!, only: [:create] 
   
-  expose :answers, ->{ Answer.all }
+  expose :answers, -> { Answer.all }
   expose :answer
-  expose :question # , -> { Question.find(params[:question_id]) }
+  expose :question #, -> { Question.find(params[:question_id]) }
 
-  def create    
-    @exposed_answer = question.answers.new(answer_params)
-    answer.author = current_user
+  def create
+    @answer = question.answers.new(answer_params)
+    @answer.author = current_user
+    @answer.save
+  end
 
-    if answer.save
-      redirect_to question, notice: 'Reply successfully sent.'
-    else
-      render 'questions/show'
-    end
+  def update
+    @question = answer.question
+    answer.update(answer_params) if current_user.author_of?(answer)
+    
+    #if current_user.author_of?(answer)
+    #  answer.update(answer_params)
+    #else
+    #  redirect_to @question, notice: 'You have no rights to do this.'
+    #end
   end
 
   def destroy
-    if current_user.author_of?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer was successfully deleted.'
+    answer.destroy if current_user.author_of?(answer)
+
+    #if current_user.author_of?(answer)
+    #  answer.destroy
+    #  # redirect_to answer.question, notice: 'Your answer was successfully deleted.'
+    #else
+    #  redirect_to answer.question, notice: 'You have no rights to do this.'
+    #end
+  end
+
+  def best
+    if current_user.author_of?(answer.question)
+      answer.make_best
     else
       redirect_to answer.question, notice: 'You have no rights to do this.'
     end
@@ -31,4 +46,5 @@ class AnswersController < ApplicationController
   def answer_params
     params.require(:answer).permit(:body)
   end
+
 end
